refactor: Simplify import of Padded
