+++ src/components/BottomButton.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Button } from 'cozy-ui/transpiled/react'
 
 import Bottom from 'components/Bottom'
-import Padded from 'components/Spacing/Padded'
+import Padded from 'components/Padded'
 import useBreakpoints from 'cozy-ui/transpiled/react/hooks/useBreakpoints'
 
 const buttonStyle = {
