+++ src/components/Padded/Padded.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import cx from 'classnames'
 
-import styles from 'components/Spacing/Padded.styl'
+import styles from 'components/Padded/Padded.styl'
 
 const _Padded = ({ className, ...restProps }) => (
   <div className={cx(styles.Padded, className)} {...restProps} />
