+++ src/components/Padded/index.js
@@ -1,0 +1,1 @@
+export { default, Unpadded } from './Padded'
