--- src/components/Spacing/index.js
@@ -1,3 +1,0 @@
-import Padded from 'components/Spacing/Padded'
-
-export { Padded }
