+++ src/ducks/balance/AccountsImporting.jsx
@@ -7,7 +7,7 @@ import { useI18n } from 'cozy-ui/transpiled/react'
 
 import Figure from 'cozy-ui/transpiled/react/Figure'
 import Header from 'components/Header'
-import { Padded } from 'components/Spacing'
+import Padded from 'components/Padded'
 import VerticalBox from 'components/VerticalBox'
 import BalancePanels from 'ducks/balance/BalancePanels'
 import BarTheme from 'ducks/bar/BarTheme'
