+++ src/ducks/loan/LoanListPage.jsx
@@ -10,7 +10,7 @@ import NarrowContent from 'cozy-ui/transpiled/react/NarrowContent'
 import AccountIcon from 'components/AccountIcon'
 import withFilters from 'components/withFilters'
 import { BalanceDetailsHeader } from 'ducks/balance'
-import { Padded } from 'components/Spacing'
+import Padded from 'components/Padded'
 import { flowRight as compose } from 'lodash'
 import { accountsConn } from 'doctypes'
 import { useBreakpoints } from 'cozy-ui/transpiled/react'
