+++ src/ducks/reimbursements/Reimbursements.jsx
@@ -6,7 +6,7 @@ import { flowRight as compose, sumBy } from 'lodash'
 import cx from 'classnames'
 import { TransactionList } from 'ducks/transactions/Transactions'
 import { translate, useI18n } from 'cozy-ui/transpiled/react'
-import { Padded } from 'components/Spacing'
+import Padded from 'components/Padded'
 import Figure from 'cozy-ui/transpiled/react/Figure'
 import styles from 'ducks/reimbursements/Reimbursements.styl'
 import Loading from 'components/Loading'
