+++ src/ducks/reimbursements/ReimbursementsPage.jsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux'
 import { useI18n } from 'cozy-ui/transpiled/react/I18n'
 import useBreakpoints from 'cozy-ui/transpiled/react/hooks/useBreakpoints'
 import Header from 'components/Header'
-import { Padded } from 'components/Spacing'
+import Padded from 'components/Padded'
 import { PageTitle } from 'components/Title'
 import cx from 'classnames'
 import BackButton from 'components/BackButton'
